% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even.odd.fun.R
\name{odd_even_fun}
\alias{odd_even_fun}
\title{Finds if the number is even or odd}
\usage{
odd_even_fun(num = 2)
}
\arguments{
\item{symbol}{}
}
\value{
null
}
\description{
The package accpets a number as an argument and checks if the number passed is even or odd
}
\examples{
even_odd_fun(2)

}
